First integration test completed
Rewrite a not_more_than_max_weekly as it is hard to remember that you are always looking for failing cases, not passing cases (backwards from regular TDD).

Base class implemented and rewrote a lot of the seed data to make more sense.

Tests should be quite quick to write. Need to redo how we run them (make test and make testpy).